Soporte para fullscreen

+++ client_src/ClientSettings.cpp
@@ -10,7 +10,7 @@ ClientSettings::ClientSettings(int screenWidth, int screenHeight, int fps, int f
 
 ClientSettings::~ClientSettings() {}
 
-// Seguramente haya que mover PI y distance a clase aparte.
-float ClientSettings::distance(float x1, float y1, float x2, float y2) {
+// Seguramente haya que mover distance a clase aparte.
+double ClientSettings::distance(double x1, double y1, double x2, double y2) {
   return sqrt(pow(x2-x1, 2) + pow(y2-y1, 2));
 }
